Added the creation of the composed time series id when inferring against a timeseries.
