+++ api/routes/grouping.go
@@ -16,7 +16,6 @@
 package routes
 
 import (
-	"fmt"
 	"net/http"
 	"strings"
 
@@ -25,11 +24,8 @@ import (
 
 	"github.com/uncharted-distil/distil-compute/model"
 	api "github.com/uncharted-distil/distil/api/model"
+	"github.com/uncharted-distil/distil/api/task"
 	"github.com/uncharted-distil/distil/api/util/json"
-)
-
-const (
-	defaultSeparator = "_"
 )
 
 // GroupingHandler generates a route handler that adds a grouping.
@@ -74,7 +70,7 @@ func GroupingHandler(dataCtor api.DataStorageCtor, metaCtor api.MetadataStorageC
 
 			if grouping.IDCol != "" {
 				// Create a new variable and column for the time series key.
-				if err := createComposedVariable(meta, data, dataset, grouping.IDCol, grouping.Properties.YCol, grouping.SubIDs); err != nil {
+				if err := task.CreateComposedVariable(meta, data, dataset, grouping.IDCol, grouping.Properties.YCol, grouping.SubIDs); err != nil {
 					handleError(w, errors.Wrapf(err, "unable to create new variable %s", grouping.IDCol))
 					return
 				}
@@ -84,7 +80,7 @@ func GroupingHandler(dataCtor api.DataStorageCtor, metaCtor api.MetadataStorageC
 			}
 
 			// Create a new grouped variable for the time series.
-			groupingVarName := strings.Join([]string{grouping.Properties.XCol, grouping.Properties.YCol}, defaultSeparator)
+			groupingVarName := strings.Join([]string{grouping.Properties.XCol, grouping.Properties.YCol}, task.DefaultSeparator)
 			err = meta.AddGroupedVariable(dataset, groupingVarName, grouping.Properties.YCol, model.TimeSeriesType, model.VarDistilRoleGrouping, grouping)
 			if err != nil {
 				handleError(w, err)
@@ -92,7 +88,7 @@ func GroupingHandler(dataCtor api.DataStorageCtor, metaCtor api.MetadataStorageC
 			}
 		} else if model.IsGeoCoordinate(grouping.Type) {
 			// No key required in this case.
-			groupingVarName := strings.Join([]string{grouping.Properties.XCol, grouping.Properties.YCol}, defaultSeparator)
+			groupingVarName := strings.Join([]string{grouping.Properties.XCol, grouping.Properties.YCol}, task.DefaultSeparator)
 			err = meta.AddGroupedVariable(dataset, groupingVarName, "Geocoordinate", model.GeoCoordinateType, model.VarDistilRoleGrouping, grouping)
 			if err != nil {
 				handleError(w, err)
@@ -177,93 +173,6 @@ func RemoveGroupingHandler(dataCtor api.DataStorageCtor, metaCtor api.MetadataSt
 			return
 		}
 	}
-}
-
-func createComposedVariable(metaStorage api.MetadataStorage, dataStorage api.DataStorage,
-	dataset string, composedVarName string, composedVarDisplayName string, sourceVarNames []string) error {
-
-	// create the variable data store entry
-	datasetStorageName := model.NormalizeDatasetID(dataset)
-
-	varExists, err := metaStorage.DoesVariableExist(dataset, composedVarName)
-	if err != nil {
-		return err
-	}
-
-	if !varExists {
-		// create the variable metadata entry
-		err := metaStorage.AddVariable(dataset, composedVarName, composedVarDisplayName, model.StringType, model.VarDistilRoleGrouping)
-		if err != nil {
-			return err
-		}
-
-		err = dataStorage.AddVariable(dataset, datasetStorageName, composedVarName, model.StringType)
-		if err != nil {
-			return err
-		}
-	}
-
-	composedData := map[string]string{}
-	var filter *api.FilterParams
-	if len(sourceVarNames) > 0 {
-		// Fetch data using the source names as the filter
-		filter = &api.FilterParams{
-			Variables: sourceVarNames,
-		}
-	} else {
-		// No grouping column - just use the d3mIndex as we'll just stick some placeholder
-		// data in.
-		filter = &api.FilterParams{
-			Variables: []string{model.D3MIndexName},
-		}
-	}
-	rawData, err := dataStorage.FetchData(dataset, datasetStorageName, filter, false)
-	if err != nil {
-		return err
-	}
-
-	// Create a map of the retreived fields to column number.  Store d3mIndex since it needs to be directly referenced
-	// further along.
-	d3mIndexFieldindex := -1
-	colNameToIdx := make(map[string]int)
-	for i, c := range rawData.Columns {
-		if c.Label == model.D3MIndexName {
-			d3mIndexFieldindex = i
-		} else {
-			colNameToIdx[c.Label] = i
-		}
-	}
-
-	if len(sourceVarNames) > 0 {
-		// Loop over the fetched data, composing each column value into a single new column value using the
-		// separator.
-		for _, r := range rawData.Values {
-			// create the hash from the specified columns
-			composed := createComposedFields(r, sourceVarNames, colNameToIdx, defaultSeparator)
-			composedData[fmt.Sprintf("%v", r[d3mIndexFieldindex].Value)] = composed
-		}
-	} else {
-		// Loop over the fetched d3mIndex values and set a placeholder value.
-		for _, r := range rawData.Values {
-			composedData[fmt.Sprintf("%v", r[d3mIndexFieldindex].Value)] = "__timeseries"
-		}
-	}
-
-	// Save the new column
-	err = dataStorage.UpdateVariableBatch(datasetStorageName, composedVarName, composedData)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func createComposedFields(data []*api.FilteredDataValue, fields []string, mappedFields map[string]int, separator string) string {
-	dataToJoin := make([]string, len(fields))
-	for i, field := range fields {
-		dataToJoin[i] = fmt.Sprintf("%v", data[mappedFields[field]].Value)
-	}
-	return strings.Join(dataToJoin, separator)
 }
 
 func setDataType(esStorage api.MetadataStorage, pgStorage api.DataStorage,
